+++ packages/mewkit/src/core/index.ts
@@ -6,6 +6,7 @@ export * from "./substitute-placeholders.js";
 export * from "./validate-install.js";
 export * from "./smart-update.js";
 export * from "./smart-update-utils.js";
+export * from "./find-orphans.js";
 export * from "./core-logger.js";
 export * from "./skills-dependencies.js";
 export * from "./dependency-installer.js";